requirements.txt
medseg~=1.0
imagehash~=4.3.1
numpy~=1.24.1
pandas~=1.5.3
scikit-image~=0.19.3
click~=8.0.4
pillow~=9.3.0
torchinfo~=1.8.0
oyaml~=1.0
tqdm~=4.64.1
setuptools~=65.6.3
torch~=2.0.0
matplotlib~=3.7.0
opencv-python~=4.7.0.72
torchvision~=0.15.1
tensorboardx~=2.6
hyperopt~=0.2.7
scikit-learn~=1.2.1
seaborn~=0.12.2
tabulate~=0.8.10
pytest~=7.1.2
beartype~=0.13.1
medpy~=0.4.0
cityscapesscripts~=2.2.2
wget~=3.2
timm~=0.9.2
einops~=0.6.0